chore: Remove unnecessary classes and enum because i dont need them anymore
